Logic

+++ src/main/java/pl/jkuznik/trucktracking/domain/trailer/TrailerController.java
@@ -4,11 +4,14 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import pl.jkuznik.trucktracking.domain.trailer.api.command.AddTrailerCommand;
+import pl.jkuznik.trucktracking.domain.trailer.api.command.UpadeteAssignmentTrailerCommand;
+import pl.jkuznik.trucktracking.domain.trailer.api.command.UpdateCrossHitchTrailerCommand;
 import pl.jkuznik.trucktracking.domain.trailer.api.dto.TrailerDTO;
 import pl.jkuznik.trucktracking.domain.truckTrailerHistory.api.dto.TruckTrailerHistoryDTO;
 
 import java.time.Instant;
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Optional;
 import java.util.UUID;
 
@@ -17,14 +20,10 @@ import java.util.UUID;
 @RequiredArgsConstructor
 public class TrailerController {
 
+    // todo czy przewidziana jest możliwość edycji numeru rejestracyjnego naczepy, a jeżeli tak to czy w bazie danych
+
+    // historia przypisań naczepy do pojazdu powinna uwzględnić poprzednią rejestrację
     private final TrailerService trailerService;
-
-    @GetMapping("/{uuid}")
-    public ResponseEntity<TrailerDTO> getTrailer(@PathVariable String uuid) {
-        TrailerDTO trailerDTO = trailerService.getTrailerByBusinessId(UUID.fromString(uuid));
-
-        return ResponseEntity.ok(trailerDTO);
-    }
 
     @GetMapping
     public ResponseEntity<List<TrailerDTO>> getTrailers() {
@@ -33,13 +32,19 @@ public class TrailerController {
         return ResponseEntity.ok(trailers);
     }
 
+    @GetMapping("/{uuid}")
+    public ResponseEntity<TrailerDTO> getTrailer(@PathVariable String uuid) {
+        TrailerDTO trailerDTO = trailerService.getTrailerByBusinessId(UUID.fromString(uuid));
+
+        return ResponseEntity.ok(trailerDTO);
+    }
+
     @GetMapping("/search")
     public ResponseEntity<List<TrailerDTO>> getTrailersActualState(
             @RequestParam(required = false) Optional<Instant> startDate,
             @RequestParam(required = false) Optional<Instant> endDate,
-            @RequestParam(required = false) Optional<Boolean> inUse,
             @RequestParam(required = false) Optional<Boolean> crossHitch) {
-        return ResponseEntity.ok(trailerService.getTrailersByFilters(startDate, endDate, inUse, crossHitch));
+        return ResponseEntity.ok(trailerService.getTrailersByFilters(startDate, endDate, crossHitch));
     }
 
     @GetMapping("/history")
@@ -67,10 +72,44 @@ public class TrailerController {
         return ResponseEntity.status(201).body(responseTrailer);
     }
 
+    @PatchMapping("/{uuid}")
+    public ResponseEntity<TrailerDTO> updateTrailerByBusinessId(@PathVariable String uuid, @RequestBody UpdateCrossHitchTrailerCommand updateCrossHitchTrailerCommand) throws Exception {
+
+        return ResponseEntity.ok(trailerService.updateTrailerByBusinessId(UUID.fromString(uuid), updateCrossHitchTrailerCommand));
+    }
+
+    @PatchMapping("/{uuid}/assign-manage")
+    public ResponseEntity<TrailerDTO> assignTrailerManage(@PathVariable String uuid, @RequestBody UpadeteAssignmentTrailerCommand upadeteAssignmentTrailerCommand) throws Exception {
+        TrailerDTO updatedTrailer = trailerService.assignTrailerManageByBusinessId(UUID.fromString(uuid), upadeteAssignmentTrailerCommand);
+
+        return ResponseEntity.status(200).body(updatedTrailer);
+    }
+
+    @PatchMapping("/{uuid}/cross-hitch")
+    public ResponseEntity<String> crossHitchTrailerByBusinessId(@PathVariable String uuid, @RequestBody UpadeteAssignmentTrailerCommand upadeteAssignmentTrailerCommand) {
+        TrailerDTO processingTrailer = trailerService.getTrailerByBusinessId(UUID.fromString(uuid));
+
+        if (!processingTrailer.isCrossHitch()) {
+            return ResponseEntity.badRequest().body("Trailer is not cross hitch operation available");
+        }
+
+        if (upadeteAssignmentTrailerCommand.truckId().isEmpty()) {
+            return ResponseEntity.badRequest().body("Truck id cannot be empty in cross hitch operation");
+        }
+
+        var result = trailerService.crossHitchOperation(UUID.fromString(uuid), upadeteAssignmentTrailerCommand);
+
+        return ResponseEntity.ok(result);
+    }
+
     @DeleteMapping("/{uuid}")
     public ResponseEntity<Void> deleteTrailer(@PathVariable String uuid) {
+
+        if (trailerService.getTrailerByBusinessId(UUID.fromString(uuid)) == null) {
+            throw new NoSuchElementException("Trailer with id " + uuid + " does not exist");
+        }
+
         trailerService.deleteTrailerByBusinessId(UUID.fromString(uuid));
-//todo dorobic warunek sprawdzajacy czy istnieje naczepa o podanym id
         return ResponseEntity.noContent().build();
     }
 }
